Fixed Testing - Moved tests that were originally in the UpdateListTest class into the MatchInteractorUnitTest     - Reflects the merging of the UpdateList class into the MatchInteractor class
