+++ app/src/androidTest/java/com/group80/uoftinder/MatchInteractorUnitTest.java
@@ -3,12 +3,25 @@ package com.group80.uoftinder;
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 
+import android.service.autofill.FieldClassification;
+
 import com.group80.uoftinder.entities.User;
 import com.group80.uoftinder.firebase.realtime.UserRealtimeDbFacade;
 
 import org.junit.Test;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+/**
+ * A MatchInteractorUnitTest class that tests the functionality of the MatchInteractor class
+ */
 public class MatchInteractorUnitTest {
+
+    /**
+     *
+     */
     @Test
     public void checkMatchListsUpdatedLocal() {
         User user1 = new User("Raghav");
@@ -20,6 +33,9 @@ public class MatchInteractorUnitTest {
         assert user2.getMatches().contains(user1.getUid());
     }
 
+    /**
+     *
+     */
     @Test
     public void checkMatchListsUpdatedRemote() {
         UserRealtimeDbFacade.getUser(
@@ -38,4 +54,64 @@ public class MatchInteractorUnitTest {
                 }
         );
     }
+
+    /**
+     * Test when the currentUser does not 'like' the displayedUser.
+     */
+    @Test
+    public void currUserNoLikeDisplayedUser() {
+        // set the currentUser and displayedUser
+        User currentUser = new User("currUser");
+        User displayedUser  = new User("displayedUser");
+        // make the currentUser's viewedList and likedList
+        List<String> viewedList = new ArrayList<>();
+        List<String> likedList = new ArrayList<>();
+
+        // add displayedUser to viewedList but not likedList
+        MatchInteractor.addToList(displayedUser, currentUser, false, viewedList, likedList);
+
+        // expected results:  displayedUser in currentUser's visited list but not liked list.
+        List<String> expectedLikedList = new ArrayList<>();
+        List<String> expectedVisitedList = new ArrayList<>(Collections.singletonList(
+                displayedUser.getUid()));
+
+        // actual results:
+        List<String> actualLikedList = currentUser.getLiked();
+        List<String> actualVisitedList = currentUser.getViewed();
+
+        // assert statements
+        assertEquals(expectedLikedList, actualLikedList);
+        assertEquals(expectedVisitedList, actualVisitedList);
+    }
+
+
+    /**
+     * Test when the currentUser does 'like' the displayedUser.
+     */
+    @Test
+    public void currUserLikeDisplayedUser() {
+        // set the currentUser and displayedUser
+        User currentUser = new User("currUser");
+        User displayedUser  = new User("displayedUser");
+        // make the currentUser's viewedList and likedList
+        List<String> viewedList = new ArrayList<>();
+        List<String> likedList = new ArrayList<>();
+        // add displayedUser to viewedList and likedList
+        MatchInteractor.addToList(displayedUser, currentUser, true, viewedList, likedList);
+
+        // expected results:  displayedUser in currentUser's liked and visited list
+        List<String> expectedLikedList = new ArrayList<>(Collections.singletonList(
+                displayedUser.getUid()));
+        List<String> expectedVisitedList = new ArrayList<>(Collections.singletonList(
+                displayedUser.getUid()));
+
+        // actual results:
+        List<String> actualLikedList = currentUser.getLiked();
+        List<String> actualVisitedList = currentUser.getViewed();
+
+        // assert statements
+        assertEquals(expectedLikedList, actualLikedList);
+        assertEquals(expectedVisitedList, actualVisitedList);
+    }
+
 }
