--- app/src/test/java/com/group80/uoftinder/UpdateListTest.java
@@ -1,71 +1,0 @@
-package com.group80.uoftinder;
-
-import static org.junit.Assert.assertEquals;
-
-import com.group80.uoftinder.entities.User;
-
-import org.junit.Test;
-
-import java.util.ArrayList;
-import java.util.List;
-import com.group80.uoftinder.entities.User;
-
-/**
- * Test class for UpdateList class.
- */
-public class UpdateListTest {
-    // create currentUser
-    User currentUser = new User("currUser");
-    // create an instance of updateList
-    UpdateList update = new UpdateList(currentUser);
-
-    @Test
-    /**
-     * Test when the currentUser does not 'like' the displayedUser.
-     */
-    public void currUserNoLikeDisplayedUser() {
-        // set the currentUser and displayedUser
-        User displayedUser  = new User("displayedUser");
-        // make the currentUser's viewedList and likedList
-        List<String> viewedList = new ArrayList<>();
-        List<String> likedList = new ArrayList<>();
-
-        // add displayedUser to viewedList but not likedList
-        UpdateList.addToList(displayedUser,false, viewedList, likedList);
-
-        // expected results:  displayedUser in currentUser's visited list but not liked list.
-        List<String> expectedLikedList = new ArrayList<>();
-        List<String> expectedVisitedList = new ArrayList<>();
-        expectedVisitedList.add(displayedUser.getUid());
-
-        // assert statements
-        assert expectedLikedList.equals(currentUser.getLiked());
-        assert expectedVisitedList.equals(currentUser.getViewed());
-    }
-
-    @Test
-    /**
-     * Test when the currentUser does 'like' the displayedUser.
-     */
-    public void currUserLikeDisplayedUser() {
-        // set the currentUser and displayedUser
-        User displayedUser  = new User("displayedUser");
-        // make the currentUser's viewedList and likedList
-        List<String> viewedList = new ArrayList<>();
-        List<String> likedList = new ArrayList<>();
-        // add displayedUser to viewedList and likedList
-        UpdateList.addToList(displayedUser,true, viewedList, likedList);
-
-        // expected results:  displayedUser in currentUser's liked and visited list
-        List<String> expectedLikedList = new ArrayList<>();
-        expectedLikedList.add(displayedUser.getUid());
-        List<String> expectedVisitedList = new ArrayList();
-        expectedVisitedList.add(displayedUser.getUid());
-
-        // assert statements
-        assert expectedLikedList.equals(currentUser.getLiked());
-        assert expectedVisitedList.equals(currentUser.getViewed());
-    }
-
-
-}
